ClimFluids

Documentation for ClimFluids.

ClimFluids.ClimFluids

ClimFluids.AbstractFluid

ClimFluids.IdealPerfectGas

ClimFluids.VarCpPerfectGas

ClimFluids.canonical_state

ClimFluids.conservative_variable

ClimFluids.exner_functions

ClimFluids.potential_enthalpy

ClimFluids.potential_temperature

ClimFluids.potential_volume

ClimFluids.pressure

ClimFluids.sound_speed

ClimFluids.sound_speed2

ClimFluids.specific_enthalpy

ClimFluids.specific_entropy

ClimFluids.specific_internal_energy

ClimFluids.specific_volume

ClimFluids.temperature

ClimFluids.ClimFluids — Module

ClimFluids provides types and functions to compute thermodynamic functions of various fluids.

Using ClimFluids

    F = Float32
    N = 1_000_000
    p = F.(1e5 .+ randn(N))
    T = F.(300. .+ randn(N));

    params = (Cp=1000.0, kappa=2/7, p0=1e5, T0=273.15)
    params = (; consvar=:temperature, map(F, params)...)
    fluid = IdealPerfectGas(params)

    s = fluid(:p,:T).specific_entropy.(p,T)
    theta = fluid(:p,:T).conservative_variable.(p,T)

In the above example, fluid is an object describing an ideal perfect gas with its physical parameters. fluid(:p,:T) is an object derived from fluid which is specialized for thermodynamics functions taking pressure and temperature as inputs. fluid(:p,:T).specific_entropy is a function-like object computing the specific entropy of the fluid given p,T. In the named tuple of parameters passed to the constructor IdealPerfectGas, consvar = :temperature indicates that potential temperature is chosen as the conservative variable. As a result the last line computes potential temperature.

See state_variables(fluid) for a list of valid inputs to fluid(...) and state_functions(fluid) for a list of valid thermodynamic functions.

Extending ClimFluids

ClimFluids is designed to be extensible. To add a new fluid, define a custom type derived from SimpleFluid, BinaryFluid, or AbstractFluid. This type should have a constructor accepting a named tuple as single input (see above example).

Thermodynamic functions defined in ClimFluids should then be specialized for this custom type. These functions accept two arguments, one for the fluid and a named tuple for the input state variables. In the above example, the callable object fluid(:p,;T).specific_entropy calls ClimFluids.specific_entropy(fluid, (p=p, T=T)). Since the same thermodynamic function can be evaluated for different tuples of state variables, thermodynamic functions must be specialized also for the type of their second argument. ClimFluids provides shortcuts for these types. For example ClimFluids.PT is the type NamedTuple{(:p,:T)}.

Thermodynamic functions may be called in performance-critical inner loops. It is recommended to make them @inline and @fastmath.

ClimFluids.AbstractFluid — Type

AbstractFluid{F,N}

Parent type for types describing thermodynamics of a fluid with N components.

ClimFluids.IdealPerfectGas — Type

IdealPerfectGas(consvar,kappa,Cp,p0,T0)

Returns an object describing the thermodynamics of a single-component ideal perfect gas. This object can then be used as first argument of thermodynamic functions. Argument consvar defines the conservative variable chosen. Legal values are :temperature, :entropy, :enthalpy, :volume, corresponding to potential temperature, entropy, potential enthalpy, potential specific volume.

ClimFluids.VarCpPerfectGas — Method

VarCpPerfectGas(kappa,Cp,p0,T0,nu)

Returns an object describing the thermodynamics of a single-component perfect gas with variable Cp. This object can then be used as first argument of thermodynamic functions.

ClimFluids.canonical_state — Method

cstate = canonical_state(fluid, state)

Converts the named tuple state of state variables to the preferred tuple cstate. It is expected that all thermodynamic functions are implemented for arguments (fluid, cstate). This function is used internally to provide fallback implementations for thermodynamic functions.

ClimFluids.conservative_variable — Function

consvar = conservative_variable(fluid, state)

Returns the conservative variable given a fluid and a state (named tuple of state variables). Which conservative variable is chosen is a parameter given to the constructor of fluid.

ClimFluids.exner_functions — Function

h, v, exner, [exner_q] = exner_functions(fluid, state)

Returns specific enthalpy, regarded as a function of pressure, conservative variable and (if applicable) composition, and its derivatives : specific volume, and Exner-like functions.

ClimFluids.potential_enthalpy — Function

hpot = potential_enthalpy(fluid, state)

Returns potential enthalpy given a fluid and a state (named tuple of state variables).

ClimFluids.potential_temperature — Function

theta = potential_temperature(fluid, state)

Returns potential temperature given a fluid and a state (named tuple of state variables).

ClimFluids.potential_volume — Function

vpot = potential_volume(fluid, state)

Returns potential volume given a fluid and a state (named tuple of state variables).

ClimFluids.pressure — Function

p = pressure(fluid, state)

Returns pressure given a fluid and a state (named tuple of state variables).

ClimFluids.sound_speed — Function

cs = sound_speed(fluid, (;p,T))

Returns the speed of sound given a fluid and a state (named tuple of state variables). Default implementation returns sqrt(sound_speed2(fluid, state)).

ClimFluids.sound_speed2 — Function

c2 = sound_speed2(fluid, (;p,T))

Returns the squared speed of sound given a fluid and a state (named tuple of state variables).

ClimFluids.specific_enthalpy — Function

h = specific_enthalpy(fluid, state)

Returns specific enthalpy given a fluid and a state (named tuple of state variables).

ClimFluids.specific_entropy — Function

s = specific_entropy(fluid, state)

Returns specific entropy given a fluid and a state (named tuple of state variables).

ClimFluids.specific_internal_energy — Function

e = specific_internal_energy(fluid, state)

Returns specific internal energy given a fluid and a state (named tuple of state variables).

ClimFluids.specific_volume — Function

v = specific_volume(fluid, state)

Returns specific volume given a fluid and a state (named tuple of state variables).

ClimFluids.temperature — Function

T = temperature(fluid, state)

Returns temperature given a fluid and a state (named tuple of state variables).