ClimFluids

Documentation for ClimFluids.

ClimFluids.AbstractFluid

ClimFluids.IdealPerfectGas

ClimFluids.VarCpPerfectGas

ClimFluids.canonical_state

ClimFluids.conservative_variable

ClimFluids.exner_functions

ClimFluids.potential_enthalpy

ClimFluids.potential_temperature

ClimFluids.potential_volume

ClimFluids.pressure

ClimFluids.sound_speed

ClimFluids.sound_speed2

ClimFluids.specific_enthalpy

ClimFluids.specific_entropy

ClimFluids.specific_internal_energy

ClimFluids.specific_volume

ClimFluids.temperature

ClimFluids.AbstractFluid — Type

AbstractFluid{F,N}

Parent type for types describing thermodynamics of a fluid with N components.

ClimFluids.IdealPerfectGas — Type

IdealPerfectGas(consvar,kappa,Cp,p0,T0)

Returns an object describing the thermodynamics of a single-component ideal perfect gas. This object can then be used as first argument of thermodynamic functions. Argument consvar defines the conservative variable chosen. Legal values are :temperature, :entropy, :enthalpy, :volume, corresponding to potential temperature, entropy, potential enthalpy, potential specific volume.

ClimFluids.VarCpPerfectGas — Method

VarCpPerfectGas(kappa,Cp,p0,T0,nu)

Returns an object describing the thermodynamics of a single-component perfect gas with variable Cp. This object can then be used as first argument of thermodynamic functions.

ClimFluids.canonical_state — Method

cstate = canonical_state(fluid, state)

Converts the named tuple state of state variables to the preferred tuple cstate. It is expected that all thermodynamic functions are implemented for arguments (fluid, cstate). This function is used internally to provide fallback implementations for thermodynamic functions.

ClimFluids.conservative_variable — Function

consvar = conservative_variable(fluid, state)

Returns the conservative variable given a fluid and a state (named tuple of state variables). Which conservative variable is chosen is a parameter given to the constructor of fluid.

ClimFluids.exner_functions — Function

h, v, exner, [exner_q] = exner_functions(fluid, state)

Returns specific enthalpy, regarded as a function of pressure, conservative variable and (if applicable) composition, and its derivatives : specific volume, and Exner-like functions.

ClimFluids.potential_enthalpy — Function

hpot = potential_enthalpy(fluid, state)

Returns potential enthalpy given a fluid and a state (named tuple of state variables).

ClimFluids.potential_temperature — Function

theta = potential_temperature(fluid, state)

Returns potential temperature given a fluid and a state (named tuple of state variables).

ClimFluids.potential_volume — Function

vpot = potential_volume(fluid, state)

Returns potential volume given a fluid and a state (named tuple of state variables).

ClimFluids.pressure — Function

p = pressure(fluid, state)

Returns pressure given a fluid and a state (named tuple of state variables).

ClimFluids.sound_speed — Function

cs = sound_speed(fluid, (;p,T))

Returns the speed of sound given a fluid and a state (named tuple of state variables). Default implementation returns sqrt(sound_speed2(fluid, state)).

ClimFluids.sound_speed2 — Function

c2 = sound_speed2(fluid, (;p,T))

Returns the squared speed of sound given a fluid and a state (named tuple of state variables).

ClimFluids.specific_enthalpy — Function

h = specific_enthalpy(fluid, state)

Returns specific enthalpy given a fluid and a state (named tuple of state variables).

ClimFluids.specific_entropy — Function

s = specific_entropy(fluid, state)

Returns specific entropy given a fluid and a state (named tuple of state variables).

ClimFluids.specific_internal_energy — Function

e = specific_internal_energy(fluid, state)

Returns specific internal energy given a fluid and a state (named tuple of state variables).

ClimFluids.specific_volume — Function

v = specific_volume(fluid, state)

Returns specific volume given a fluid and a state (named tuple of state variables).

ClimFluids.temperature — Function

T = temperature(fluid, state)

Returns temperature given a fluid and a state (named tuple of state variables).